Remove extra space
I accidentally added a space at the end in the translation credit. I was worried that it would cause issues with text formatting so I have removed it. Also I wrote in the issue I posted that I would create a  pull request to fix it, but I then noticed I could propose an edit it like this. Sorry for the confusion, I'm still learning :)